+++ src/test/java/picocli/SplitSynopsisLabelTest.java
@@ -1,7 +1,6 @@
 package picocli;
 
 import picocli.CommandLine.Command;
-import picocli.CommandLine.Model.OptionSpec;
 import picocli.CommandLine.Option;
 
 import org.junit.Test;
@@ -9,12 +8,12 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static picocli.CommandLine.Help.Ansi.OFF;
 
-public class UsageSplitExecutionTest {
+public class SplitSynopsisLabelTest {
     @Test
     public void testSimple() {
         @Command(name = "UsageSplit")
         class Simple {
-            @Option(names = "v", split = "\\|", usageSplit = "|")
+            @Option(names = "v", split = "\\|", splitSynopsisLabel = "|")
             String args[] = {};
         }